Adjust load_nodes_from JSON that it also handles lists correctly (#426)
